refactor: move uploadFileUint8Array function to upload-utils and update imports in NetworkFacade and upload modules
